Moved get_ltd_systems' api requests to api_requests
Instead of pinging the api straight from cache I moved it to api_requests similarly to how the other api requests are done. I also added an if statement in the get_ltd_systems function to check the number of states before checking exactly what they are which should boost efficiency.