Added logging & added commas I forgot
In cache.py I forgot commas in the distFromBoran line (277 & 303). Added logging is a pretty big one. Created a new log.py and initialize the log in bot.py and added env variable for LOG for if you want the log to happen. Also left a spot to manually type in a path for the output of the log.  Then everywhere wehere a debug is printed I've added a statement to add to the log if that is so choosen.